backend/requirements.txt
Flask==3.1.2
Flask-CORS==5.0.0
python-dotenv==1.1.1
gunicorn==20.1.0
SQLAlchemy==2.0.36
Flask-Migrate==4.0.7
psycopg2-binary==2.9.9
plaid-python==20.0.0
PyJWT==2.8.0

